Extract interface for exception factory
