+++ plugin/src/main/groovy/com/novoda/buildproperties/internal/DefaultExceptionFactory.groovy
@@ -1,0 +1,25 @@
+package com.novoda.buildproperties.internal
+
+class DefaultExceptionFactory implements ExceptionFactory {
+
+    final String additionalMessage
+    final ConsoleRenderer consoleRenderer = new ConsoleRenderer()
+
+    DefaultExceptionFactory(String additionalMessage) {
+        this.additionalMessage = additionalMessage
+    }
+
+    @Override
+    Exception fileNotFound(File file) {
+        new FileNotFoundException("File ${consoleRenderer.asClickableFileUrl(file)} does not exist.${formattedAdditionalMessage()}")
+    }
+
+    @Override
+    Exception propertyNotFound(String key) {
+        new PropertyNotFoundException(key)
+    }
+
+    private String formattedAdditionalMessage() {
+        additionalMessage ? "\n$additionalMessage" : ''
+    }
+}
